requirements.txt
# Core deep learning
torch>=2.1.0
torchvision>=0.16.0

# Qwen VL model support
transformers>=4.49.0
accelerate>=0.26.0
qwen-vl-utils>=0.0.8

# 4-bit quantization
bitsandbytes>=0.43.0

# Video / image processing
opencv-python>=4.8.0
Pillow>=10.0.0
decord>=0.6.0

# Data handling
numpy>=1.24.0

# Evaluation metrics
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.0